Allow env update on task re-run
This is regarding the use case where a workflow execution failed
because of environment related issues (i.e. endpoint unavailable,
etc.). Endpoints and credentials can be passed on workflow invocation
under the env param and then accessed by workflow tasks using the
env() function. It is possible that the endpoint is replaced (i.e.
different host/ip) as part of resolving the environment related
issues or that the token passed as credential has expired. In these
cases, the user will need to be able to update the env variables
prior to re-run.

Change-Id: Icc8faf990a3cbdb8e969ae4cda56c0d06f11155f