Add a policy section to the Mistral specs
The idea of policies is to document something related to the project
that isn't tied to a specific release and often isn't related to code.
For example, a policy could be related to how we organise or how we
agree to do code reviews.

This process is inspired by the policies used in TripleO and the policy
template is a direct copy of the one used in TripleO.

Change-Id: I05f5076a1559790bf6d0f7b68d528748a2acd20b