Create and run workflows within a namespace
Creating and running workflows within a namespace will allow users to create
many workflows with the same name. This is useful when a user already has many
workflows that are connected to each other implemented and one of the workflow
names is already in use and the user does not want to edit the that workflow
and all the ones referencing it or merge them to a workbook.

blueprint create-and-run-workflows-within-a-namespace

Change-Id: I5a8e6c8761bfe31bf04026083c4755468f580429